Found the buggit status now just need to fix the test
Problem was this: Once the decipher.final() method has been called, the Decipher object can no longer be used to decrypt data. Attempts to call decipher.final() more than once will result in an error being thrown. the deciper object was created at the start of the script and was trying to be used multiple times, but once the final() method was called the object was destryed

ref:https://nodejs.org/api/crypto.html#crypto_decipher_final_output_encoding